+++ services/src/pce/src/main/java/org/openkilda/pce/provider/NeoDriver.java
@@ -307,7 +307,7 @@ public class NeoDriver implements PathComputer, TopologyRepository {
     }
 
     @Override
-    public AvailableNetwork getAvailableNetwork(boolean ignoreBandwidth, int requestedBandwidth) {
+    public AvailableNetwork getAvailableNetwork(boolean ignoreBandwidth, long requestedBandwidth) {
         return new AvailableNetwork(driver, ignoreBandwidth, requestedBandwidth);
     }
